Refactor -> move socket config into its own file
