+++ server/src/server.js
@@ -1,8 +1,7 @@
 import app from "./app.js";
-import { corsOptions } from "./configuration/corsOptions.js";
 import { connectToDatabase } from "./database/connectToDB.js";
 import { createServer } from "http";
-import { Server as SocketIOServer } from "socket.io";
+import { initSockets } from "./sockets/socket.js";
 
 export function initServer() {
   // Connect to DB
@@ -11,27 +10,8 @@ export function initServer() {
   // Crée un serveur HTTP avec Express
   const httpServer = createServer(app);
 
-  // Initialise Socket.io sur ce serveur
-  const io = new SocketIOServer(httpServer, {
-    cors: corsOptions,
-  });
-  const userSocketMap = {};
-
-  io.on("connection", (socket) => {
-    const userId = socket.handshake.query.userId;
-
-    if (userId) {
-      userSocketMap[userId] = socket.id;
-      io.emit("getOnlineUsers", Object.keys(userSocketMap));
-    }
-
-    socket.on("disconnect", () => {
-      if (userId) {
-        delete userSocketMap[userId];
-        io.emit("getOnlineUsers", Object.keys(userSocketMap));
-      }
-    });
-  });
+  // Initialise les WebSockets
+  initSockets(httpServer);
 
   if (!process.env.PORT) {
     console.error("Please specify the port number for the HTTP server with the environment variable PORT in the .env file.");
